feat: create PlaylistsList and UserSearchBox components
